// VIS_Application/app/Model/Notification/HelpTicketAdd.ts
export interface IHelpAddTicket {
    CompanyId: number;
    Id: number;
    Subject: string;
    Message: string;
    Remarks: string;
    AssignTo: number;
    AddressToGroup: number;
    Priority: string
    Status: string;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;

}